KAFKA-20297 Move ChildFirstClassLoader, ProducerIdAndEpoch into internal
We have exposed several classes in org.apache.kafka.common.utils as
public APIs, which can be misleading since not all classes in this
package are intended for external use.

So move following class into internal module
- ChildFirstClassLoader
- ProducerIdAndEpoch
