+++ core/src/test/scala/unit/kafka/coordinator/transaction/TransactionCoordinatorConcurrencyTest.scala
@@ -35,8 +35,9 @@ import org.apache.kafka.common.network.ListenerName
 import org.apache.kafka.common.protocol.{ApiKeys, Errors}
 import org.apache.kafka.common.record.internal.{FileRecords, MemoryRecords, RecordBatch, SimpleRecord}
 import org.apache.kafka.common.requests._
-import org.apache.kafka.common.utils.{MockTime, ProducerIdAndEpoch}
+import org.apache.kafka.common.utils.MockTime
 import org.apache.kafka.common.utils.internals.LogContext
+import org.apache.kafka.common.utils.internals.ProducerIdAndEpoch
 import org.apache.kafka.common.{Node, TopicPartition, Uuid}
 import org.apache.kafka.coordinator.transaction.{InitProducerIdResult, ProducerIdManager, TransactionConfig, TransactionLog, TransactionMetadata, TransactionState}
 import org.apache.kafka.metadata.MetadataCache
